requirements.txt
geopandas>=0.12.2
shapely>=2.0.1
pandas>=1.5.3
numpy>=1.24.2
geopy>=2.3.0
tqdm>=4.64.1
joblib>=1.2.0
psycopg2-binary>=2.9.6